tracker/requirements.txt
beautifulsoup4>=4.12.0
requests>=2.31.0
selenium>=4.15.0
playwright>=1.40.0
python-dotenv>=1.0.0
schedule>=1.2.0
sqlalchemy>=2.0.0
boto3>=1.28.0
pynamodb>=5.5.0
